LayoutTests/http/tests/security/contentSecurityPolicy/script-src-strict-dynamic-and-script-src-elem-expected.txt

ALERT: importScripts allowed
ALERT: importScripts allowed
